Migrate Avalanche consensus into core with a shared base (0.6.0).
Move Slush/Snowflake/Snowball into chaincraft/consensus/gossip/avalanche.py
under a common BinarySamplingConsensus base, removing ~200 lines of duplication
per protocol. The example modules become thin re-export shims; all existing
unit and network integration tests pass unchanged.
